requirements.txt
Python ≥ 3.7
PyTorch
torchvision
timm
scikit-learn
matplotlib
pillow